requirements.txt
# pytest==3.9.1
# flake8==3.5.0
numpy==1.22.3
tensorflow==2.7.0
scipy==1.7.3
Pillow==8.3.1
requests==2.24.0
googledrivedownloader==0.3
tqdm==4.28.0
aws-lambda-typing==2.10.1
boto3==1.21.44